pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent"
version = "0.1.0"
description = "Local CLI coding agent powered by Ollama"
requires-python = ">=3.9"
dependencies = [
    "ollama>=0.3.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "gitpython>=3.1.40",
    "mcp>=1.0.0",
]

[project.scripts]
coding-agent = "agent.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
